him-api/requirements.txt
gunicorn==20.1.0
psycopg2==2.9.5
pyyaml==6.0
google-cloud-secret-manager==2.12.6
google-auth==2.13.0
requests==2.28.1
openai==0.25.0
djangorestframework==3.14.0
django-filter==22.1
django-extensions==3.2.1
django-environ==0.9.0
django-storages[google]==1.13.1
pyyaml==6.0
djangorestframework-simplejwt==5.2.2